+++ com.unity.shadergraph/Editor/Data/Nodes/Math/Derivative/DDYNode.cs
@@ -14,18 +14,19 @@ namespace UnityEditor.ShaderGraph
 
         protected override MethodInfo GetFunctionToConvert()
         {
-            return GetType().GetMethod("Unity_DDY", BindingFlags.Static | BindingFlags.NonPublic);
+            return GetType().GetMethod("Unity_DDY", BindingFlags.Instance | BindingFlags.NonPublic);
         }
 
-        static string Unity_DDY(
+        string Unity_DDY(
             [Slot(0, Binding.None)] DynamicDimensionVector In,
             [Slot(1, Binding.None, ShaderStageCapability.Fragment)] out DynamicDimensionVector Out)
         {
             return
-@"
-{
+$@"
+{{
+    {GetRayTracingError()}
     Out = ddy(In);
-}
+}}
 ";
         }
     }